requirements.txt
ultralytics
yt-dlp
opencv-python
pandas
google-genai
python-dotenv
ipykernel
pillow
